Fix cartelle vuote + miglioramento log

+++ client/src/file/SyncFileWatcher.cpp
@@ -8,24 +8,32 @@
 #include "SyncFileWatcher.h"
 #include "common/hash_file.h"
 
-SyncFileWatcher::SyncFileWatcher(FileWatcher &fw, ClientCommand &c): fw{fw}, c{c} {}
+
+SyncFileWatcher::SyncFileWatcher(FileWatcher &fw, ClientCommand &c): fw{fw}, c{c} {
+    for(auto &file : boost::filesystem::recursive_directory_iterator(fw.get_path_to_watch())) {
+        if (boost::filesystem::is_directory(file)) {
+            directories.insert(file.path().string());
+        }
+    }
+}
 
 // Questa funzione setta la callback che gestisce le azioni necessarie in base al tipo di evento
 // (modifica, creazione o cancellazione) avvenuto.
 void SyncFileWatcher::run() {
     this->fw.on_file_changed([this](std::string path_matched, FileStatus status) -> void {
-        if (!boost::filesystem::is_regular_file(boost::filesystem::path(path_matched)) && status != FileStatus::erased) {
-            return;
-        }
         FileMetadata fm;
         fm.path = path_matched;
         if (fm.path.find("/.") != std::string::npos) return;
 
-        std::size_t found = fw.path_to_watch.string().find_last_of("/\\");
+        std::size_t found = fw.get_path_to_watch().string().find_last_of("/\\");
         fm.path_to_send = fm.path.substr(found);
 
 
         if (status == FileStatus::created || status == FileStatus::modified) {
+            if (boost::filesystem::is_directory(boost::filesystem::path(path_matched))) {
+                directories.insert(path_matched);
+                return;
+            }
             fm.hash = hash_file(fm.path);
             auto post_file = c.post_file(fm);
             bool post_file_result = post_file.get();
@@ -38,8 +46,14 @@ void SyncFileWatcher::run() {
                 std::cout << " ? " << fm.path << std::endl;
             }
         } else if (status == FileStatus::erased) {
+            if (directories.count(path_matched)) {
+                directories.erase(path_matched);
+                return;
+            }
+
             auto remove_file = c.remove_file(fm);
             bool remove_file_result = remove_file.get();
+
             if (remove_file_result) {
                 std::cout << "I seguenti file sono stati eliminati dal server: " << std::endl;
                 std::cout << " - " << fm.path << std::endl;
@@ -53,5 +67,3 @@ void SyncFileWatcher::run() {
 
     });
 }
-
-
